Add generic data migration script
This script pulls all data from one API, optionally runs migrations on
it, and pushes it to another API.

Basic usage:

    $ python migrate.py \
    > -s "https://oerworldmap.org/resource/" \
    > -t "http://localhost:9000/resource/"

To migrate data during the process, create a migration script in the
`scripts/migrations/` directory:

    # scripts/migrations/do_something.py
    def migrate(resource):
        resource['foo'] = 'bar'
        return resource

Then pass this migration script as a parameter:

    $ python migrate.py \
    > -s "https://oerworldmap.org/resource/" \
    > -t "http://localhost:9000/resource/" \
    > -m do_something

You can pass in multiple migration scripts by repeating the `-m`
parameter. They will be executed in the order they are provided.