Parially align tests with new repo setup
There is a lot to be done here to adapt the test cases. Most of it has to do
with the fact that currently, `@context` is automatically added to all
resources (even embedded ones) and the fact that a lot of test data does not
use valid URIs as `@id` which is now necessary since we have the triple store
in there. Maybe we should refactor the method signatures accordingly.

/cc @philboeselager