Fix implicitly adding embedded resources
Adding a resource with an embedded resource such as

    {
      "@id": "info:123",
      "@type": "Person",
      "name": "Alice",
      "knows": {
        "@id": "info:456",
        "@type": "Person",
        "name": "Bob"
      }
    }

will now not result in any triples with `info:456` as their subject being
added (except for potential inverse relations). Instead, a [Concise Bounded
Description](https://www.w3.org/Submission/CBD/) of the root resource is
extracted. Implicitly adding embedded resources results in overly large diffs
because the base that is compared to is always a CBD.