Made changes in the table
I inserted <th> for the head of the table and I tried to define the text alignment in the cells, left in the head and top in the other cells. In the preview only the head is working. Maybe the text align works on production. Otherwise tell me how to define the text alignment in the table.